+++ tests/framework/db/ActiveRecordTest.php
@@ -1509,7 +1509,7 @@ abstract class ActiveRecordTest extends DatabaseTestCase
     }
 
     /**
-     * Ensure no ambiguous colum error occurs if ActiveQuery adds a JOIN
+     * Ensure no ambiguous column error occurs if ActiveQuery adds a JOIN
      * https://github.com/yiisoft/yii2/issues/13757
      */
     public function testAmbiguousColumnFindOne()
@@ -1518,5 +1518,34 @@ abstract class ActiveRecordTest extends DatabaseTestCase
         $model = Customer::findOne(1);
         $this->assertTrue($model->refresh());
         CustomerQuery::$joinWithProfile = false;
+    }
+
+    /**
+     * Ensure no ambiguous column error occurs on indexBy with JOIN
+     * https://github.com/yiisoft/yii2/issues/13859
+     */
+    public function testAmbiguousColumnIndexBy()
+    {
+        switch ($this->driverName)
+        {
+            case 'pgsql':
+            case 'sqlite':
+                $selectExpression = "(customer.name || ' in ' || p.description) AS name";
+                break;
+            case 'cubird':
+            case 'mysql':
+                $selectExpression = "concat(customer.name,' in ', p.description) name";
+                break;
+            default:
+                $this->markTestIncomplete('CONCAT syntax for this DBMS is not added to the test yet.');
+        }
+
+        $result = Customer::find()->select([$selectExpression])
+            ->innerJoinWith('profile p')
+            ->indexBy('id')->column();
+        $this->assertEquals([
+            1 => 'user1 in profile customer 1',
+            3 => 'user3 in profile customer 3',
+        ], $result);
     }
 }